+++ src/config.rs
@@ -1,5 +1,5 @@
 use std::sync::Arc;
-use std::time::Duration;
+use std::time::{Duration, Instant};
 
 use crate::config::DtlsCert;
 use crate::crypto::CryptoProvider;
@@ -12,11 +12,12 @@ use crate::Rtc;
 ///
 /// ```
 /// # #[cfg(feature = "openssl")] {
+/// use std::time::Instant;
 /// use str0m::RtcConfig;
 ///
 /// let rtc = RtcConfig::new()
 ///     .set_ice_lite(true)
-///     .build();
+///     .build(Instant::now());
 /// # }
 /// ```
 ///
@@ -194,13 +195,14 @@ impl RtcConfig {
     ///
     /// ```
     /// # #[cfg(feature = "openssl")] {
+    /// # use std::time::Instant;
     /// # use str0m::RtcConfig;
     /// // For the session to use only OPUS and VP8.
     /// let mut rtc = RtcConfig::default()
     ///     .clear_codecs()
     ///     .enable_opus(true)
     ///     .enable_vp8(true)
-    ///     .build();
+    ///     .build(Instant::now());
     /// # }
     /// ```
     pub fn clear_codecs(mut self) -> Self {
@@ -531,8 +533,8 @@ impl RtcConfig {
     }
 
     /// Create a [`Rtc`] from the configuration.
-    pub fn build(self) -> Rtc {
-        Rtc::new_from_config(self).expect("Failed to create Rtc from config")
+    pub fn build(self, start: Instant) -> Rtc {
+        Rtc::new_from_config(self, start).expect("Failed to create Rtc from config")
     }
 }